Cut/copy paste ReadWrite controller code for once
Setting for to have a Once controller.
 - Copy ReadWrite controller to once.go
 - Cut ReadWrite tests that are related to once to once_test.go
